+++ go-client-sdk/sdk.go
@@ -86,7 +86,7 @@ func New(opts ...SDKOption) *SDK {
 	sdk := &SDK{
 		_language:   "go",
 		_sdkVersion: "0.5.0",
-		_genVersion: "2.28.0",
+		_genVersion: "2.29.0",
 	}
 	for _, opt := range opts {
 		opt(sdk)