+++ java-client-sdk/lib/src/main/java/industries/dot/brease/SDK.java
@@ -36,7 +36,7 @@ public class SDK {
 	private String _serverUrl;
 	private String _language = "java";
 	private String _sdkVersion = "0.3.0";
-	private String _genVersion = "2.28.0";
+	private String _genVersion = "2.29.0";
 	/**
 	 * The Builder class allows the configuration of a new instance of the SDK.
 	 */