+++ RELEASES.md
@@ -151,3 +151,13 @@ Based on:
 ### Releases
 - [Go v0.5.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.5.0 - go-client-sdk
 - [NPM v0.2.0] https://www.npmjs.com/package/brease-sdk/v/0.2.0 - typescript-client-sdk
+
+## 2023-05-17 00:43:52
+### Changes
+Based on:
+- OpenAPI Doc 0.1.0 https://prod.speakeasyapi.dev/v1/apis/brease/version/0.1/schema/download
+- Speakeasy CLI 1.33.2 (2.29.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [Go v0.5.0] https://github.com/speakeasy-sdks/brease/releases/tag/go-client-sdk/v0.5.0 - go-client-sdk
+- [PyPI v1.2.0] https://pypi.org/project/brease-sdk/1.2.0 - python-client-sdk
+- [NPM v0.3.0] https://www.npmjs.com/package/brease-sdk/v/0.3.0 - typescript-client-sdk